Create configuration option InitBIOS
Add a new configuration option: "InitBIOS". This configuration options specifies whether the card is soft-booted using the Oak video BIOS. By default it is set to "on". When disabling soft-boot, the card is left in its power-up state. This option is meant for debugging and discovering required configuration for writing the KMS driver, unable to execute the video BIOS from kernel mode.

The setting has no effect if UseBIOS is set to "on".